Move epoch from secure to nano_lib
